requirements.txt
psutil
py-cpuinfo
pySMART
tensorflow
tensorflow_quantum
cirq
qiskit
pennylane